Extracted all non-cookie-specific stuff from CookiesFieldEditorDialog
